some bugfixes + tests
