+++ polog/core/utils/is_json.py
@@ -1,9 +1,12 @@
+import json
+
+
 def is_json(value):
     """
     Функция, проверяющая, является ли переданное ей значение валидной строкой формата json.
     """
     try:
         json.loads(value)
-    except:
+    except Exception:
         return False
     return True
