+++ polog/field.py
@@ -77,7 +77,8 @@ class field:
 
     def standart_converter(self, value):
         """
-        Тупо берем любой объект и скармливаем его стандартной функции str().
+        Тупо возвращаем то, что сюда передано.
+        
         Данный метод используется, если пользователь не передал в конструктор класса кастомную функцию-сериализатор.
         """
-        return str(value)
+        return value
