+++ polog/loggers/handle/abstract.py
@@ -51,9 +51,6 @@ class AbstractHandleLogger:
     _position_args = {
         0: 'message',
     }
-    # Ключи - названия полей, значения - функции.
-    # Каждая из этих функций должна принимать словарь с уже ранее извлеченными значениями полей и возвращать значение поля, название которого является ключом.
-    _default_values = {}
     # Множество названий полей, которые запрещено логировать.
     # Данное множество просматривается в приоритете по сравнению с разрешенными полями и блокирует запись в том числе запрещенных позиционных аргументов.
     _forbidden_fields = set()
@@ -79,7 +76,7 @@ class AbstractHandleLogger:
 
         1. Извлечение данных из переданных пользователем аргументов (каких угодно) в словарь. Попутно данные проверяются на соответствие ожидаемым форматам, и, возможно, как-то еще дополнительно обрабатываются.
         2. Применение каких-то специфических методов к полученному словарю с данными для его преобразования. Набор этих методов определяется в классе-наследнике, их может вообще не быть.
-        3. Передача полученного словаря с данными куда-то, где с ним будет сделано что-то. Конкретное действие определяется в классе-наследнике.
+        4. Передача полученного словаря с данными куда-то, где с ним будет сделано что-то. Конкретное действие определяется в классе-наследнике.
         """
         fields = self._prepare_data(args, kwargs)
         self._specific_processing(fields)
@@ -114,17 +111,7 @@ class AbstractHandleLogger:
         fields = {}
         self._position_args_to_dict(args, fields, self._position_args)
         self._kwargs_to_dict(kwargs, fields)
-        self._defaults_to_dict(fields)
         return fields
-
-    def _defaults_to_dict(self, fields):
-        """
-        Некоторые значения являются обязательными, но от пользователя не поступили. В этом случае мы генерируем их автоматически.
-        В словаре self._default_values по ключам в виде названий полей лога хранятся функции. Каждая такая функция должна принимать словарь с ранее уже извлеченными полями лога и возвращать содержимое обязательного поля.
-        """
-        for key, get_default in self._default_values.items():
-            if key not in fields:
-                fields[key] = get_default(fields)
 
     def _kwargs_to_dict(self, kwargs, destination, raise_if_collision=True):
         """
